requirements.txt
bottle==0.12.19
bs4==0.0.1
tinydb==4.7.0
sanitize-filename

behave
behave-html-formatter
pyhamcrest
selenium
WebTest
gunicorn
pyjwt
pymongo
dnspython
